+++ src/main/java/com/supermartijn642/wormhole/Wormhole.java
@@ -176,5 +176,6 @@ public class Wormhole {
         handler.addGenerator(WormholeLootTableGenerator::new);
         handler.addGenerator(WormholeRecipeGenerator::new);
         handler.addGenerator(WormholeTagGenerator::new);
+        handler.addGenerator(WormholeAtlasSourceGenerator::new);
     }
 }