+++ src/main/java/com/supermartijn642/wormhole/data/WormholeBlockStateGenerator.java
@@ -48,7 +48,7 @@ public class WormholeBlockStateGenerator extends BlockStateGenerator {
         // Coal generator
         this.blockState(Wormhole.coal_generator).variantsForAll((state, variant) -> {
             boolean lit = state.get(CoalGeneratorBlock.LIT);
-            int rotation = (int)-state.get(CoalGeneratorBlock.FACING).toYRot() % 360;
+            int rotation = (180 - (int)state.get(CoalGeneratorBlock.FACING).toYRot()) % 360;
             variant.model(lit ? "block/coal_generator_lit" : "block/coal_generator", 0, rotation);
         });
     }