+++ src/main/java/com/supermartijn642/wormhole/portal/screen/ScreenBlockRenderer.java
@@ -4,7 +4,6 @@ import com.mojang.blaze3d.platform.Lighting;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.VertexConsumer;
-import com.mojang.math.Quaternion;
 import com.supermartijn642.core.ClientUtils;
 import com.supermartijn642.core.render.RenderUtils;
 import net.minecraft.client.renderer.LightTexture;
@@ -18,6 +17,7 @@ import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.client.model.data.ModelData;
+import org.joml.Quaternionf;
 
 import java.util.List;
 
@@ -36,7 +36,8 @@ public class ScreenBlockRenderer {
         MultiBufferSource.BufferSource bufferSource = RenderUtils.getMainBufferSource();
         Lighting.setupForFlatItems();
 
-        poseStack.mulPose(new Quaternion(pitch, yaw, 0, true));
+        poseStack.mulPose(new Quaternionf().setAngleAxis(pitch / 180 * Math.PI, 1, 0, 0));
+        poseStack.mulPose(new Quaternionf().setAngleAxis(yaw / 180 * Math.PI, 0, 1, 0));
 
         BakedModel model = ClientUtils.getBlockRenderer().getBlockModel(state);
         ModelData modelData = ModelData.EMPTY;